+++ example/lib/inherited/base_value.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/cupertino.dart';
+
+class BaseValue extends ChangeNotifier {
+  int count = 0;
+
+  void updateCount(int value) {
+    count = value;
+    notifyListeners();
+  }
+}